backend/drone_detection/requirements.txt
tensorflow
numpy
matplotlib
pillow
ipykernel
scikit-learn